requirements.txt
flask
pandas
numpy
scikit-learn
httpx
requests
anthropic
sqlite3
python-dotenv
reportlab
